Move existing pipeline into lang module
