+++ src/lang/mod.rs
@@ -1,0 +1,2 @@
+
+pub mod en;
